ai-service-python/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
pydantic>=2.12.0
python-dotenv==1.0.1
requests==2.32.3
numpy==2.2.3
pandas==2.2.3
scikit-learn==1.6.1
pdfplumber==0.11.5
pypdf==5.3.0
pytesseract==0.3.13
Pillow==11.1.0
reportlab==4.3.1
python-multipart==0.0.20
